+++ src/Custard/ScenarioDataSeeder.php
@@ -6,15 +6,37 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 abstract class ScenarioDataSeeder implements DemoDataSeederInterface
 {
+    private static $scenarioCount = 1;
+
+    private static $alreadyRan = [];
+
+    /**
+     * @return array|DemoDataSeederInterface[]
+     */
+    protected function getPreRequisiteSeeders()
+    {
+        return [];
+    }
+
     public function seedData(OutputInterface $output)
     {
-        $scenarioNumber = 1;
+        $class = get_class($this);
+        if (in_array($class, self::$alreadyRan)){
+            return;
+        }
+
+        self::$alreadyRan[] = $class;
+
+        foreach($this->getPreRequisiteSeeders() as $seeder){
+            $seeder->seedData($output);
+        }
+
         foreach ($this->getScenarios() as $scenario) {
             $output->writeln("");
-            $output->writeln("Scenario $scenarioNumber: " . $scenario->getName());
+            $output->writeln("Scenario ".self::$scenarioCount.": " . $scenario->getName());
             $output->writeln("");
             $scenario->run($output);
-            $scenarioNumber++;
+            self::$scenarioCount++;
         }
     }